requirements.txt
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
cfgv==3.2.0
click==7.1.2
dataclasses==0.8
distlib==0.3.1
filelock==3.0.12
identify==1.5.12
importlib-metadata==3.3.0
importlib-resources==5.0.0
iniconfig==1.1.1
mypy-extensions==0.4.3
nodeenv==1.5.0
packaging==20.8
pathspec==0.8.1
pkg-resources==0.0.0
pluggy==0.13.1
pre-commit==2.9.3
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
python-dotenv==0.15.0
PyYAML==5.3.1
regex==2020.11.13
selenium==3.141.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.3.0
zipp==3.4.0
